+++ Dockerfile
@@ -2,14 +2,10 @@ FROM python:3.13-slim-bookworm
 
 WORKDIR /app
 
-# Install runtime deps
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
+COPY pyproject.toml README.md ./
+COPY schoonmaker/ ./schoonmaker/
+RUN pip install --no-cache-dir .
 
-# Application and default sample(s)
-COPY schoonmaker/ schoonmaker/
-COPY cli.py test.py ./
-COPY samples/ samples/
+COPY samples/ ./samples/
 
-# Default: parse default sample and print summary
-CMD ["python", "cli.py", "run"]
+CMD ["schoonmaker", "run"]
