+++ pyproject.toml
@@ -1,0 +1,21 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "schoonmaker"
+description = "Parse Final Draft FDX screenplays to JSON AST or Fountain"
+readme = "README.md"
+requires-python = ">=3.9"
+dynamic = ["version"]
+dependencies = []
+
+[project.scripts]
+schoonmaker = "schoonmaker.cli:main"
+
+[tool.setuptools.dynamic]
+version = {attr = "schoonmaker.version.version"}
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["schoonmaker*"]
