+++ schoonmaker/version.py
@@ -1,1 +1,1 @@
-version = "1.0.1"
+version = "1.1.0"
